+++ ikea-zigbee-drivers/Ikea_WS-Light.groovy
@@ -35,7 +35,7 @@ metadata {
         capability 'HealthCheck'
         capability 'PowerSource'
 
-        fingerprint profileId:'0104', endpointId:'01', inClusters:'0000,0003,0004,0005,0006,0008,0300,1000,FC57', outClusters:'0019', model:'TRADFRI bulb E14 WS globe 470lm', manufacturer:'IKEA of Sweden'  // Type LED2101G4: 1.1.003 (117C-2204-00011003)
+        fingerprint profileId:'0104', endpointId:'01', inClusters:'0000,0003,0004,0005,0006,0008,0300,1000,FC57', outClusters:'0019', model:'TRADFRI bulb E14 WS globe 470lm', manufacturer:'IKEA of Sweden'  // Type LED2101G4, LED1949C5: 1.1.003 (117C-2204-00011003)
         
         // Attributes for capability.HealthCheck
         attribute 'healthStatus', 'enum', ['offline', 'online', 'unknown']
@@ -413,10 +413,8 @@ void refresh(boolean auto = false) {
     cmds += zigbee.readAttribute(0x0006, 0x4003) // PowerOnBehavior
     
     // Refresh for capability.ColorTemperature
-    cmds += zigbee.readAttribute(0x0300, 0x0008) // ColorMode
-    cmds += zigbee.readAttribute(0x0300, 0x0007) // ColorTemperatureMireds
-    cmds += zigbee.readAttribute(0x0300, 0x400B) // ColorTemperaturePhysicalMinMireds
-    cmds += zigbee.readAttribute(0x0300, 0x400C) // ColorTemperaturePhysicalMaxMireds
+    cmds += zigbee.readAttribute(0x0300, [0x0007, 0x0008]) // ColorTemperatureMireds, ColorMode
+    cmds += zigbee.readAttribute(0x0300, [0x400B, 0x400C]) // ColorTemperaturePhysicalMinMireds, ColorTemperaturePhysicalMaxMireds
     
     // Refresh for capability.Brightness
     cmds += zigbee.readAttribute(0x0008, 0x0000) // CurrentLevel
@@ -477,6 +475,41 @@ void shiftColorTemperature(String direction) {
     Integer stepSize = (state.maxMireds - state.minMireds) * Integer.parseInt(colorTemperatureStep) / 100
     String payload = "${utils_payload mode, 2} ${utils_payload stepSize, 4} 0000 ${utils_payload state.minMireds, 4} ${utils_payload state.maxMireds, 4} 00 00"
     utils_sendZigbeeCommands(["he raw 0x${device.deviceNetworkId} 0x01 0x${device.endpointId} 0x0300 {11434C ${payload}}"])
+}
+private void processMultipleColorTemperatureAttributes(Map msg, String type) {
+    Map<Integer, String> attributes = [:]
+    attributes[msg.attrInt] = msg.value
+    msg.additionalAttrs?.each { attributes[Integer.parseInt(it.attrId, 16)] = it.value }
+
+    Integer mireds = -1
+    Integer temperature = -1
+    String colorMode = null
+    String colorName = null
+    attributes.each {
+        switch (it.key) {
+            case 0x0007:
+                mireds = Integer.parseInt it.value, 16
+                temperature = Math.round(1000000 / mireds)
+                break
+
+            case 0x0008:
+            case 0x4001:
+                colorMode = it.value == '02' ? 'CT' : 'RGB'
+                utils_sendEvent name:'colorMode', value:colorMode, descriptionText:"Color mode is ${colorMode}", type:type
+                break
+        }
+    }
+
+    if (temperature >= 0) utils_sendEvent name:'colorTemperature', value:temperature, descriptionText:"Color temperature is ${temperature}K", type:type
+
+    // Update colorName, if the case
+    if ("${colorMode ?: device.currentValue('colorMode', true)}" == 'CT') {
+        Integer colorTemperature = temperature >= 0 ? temperature : device.currentValue('colorTemperature', true)
+        colorName = convertTemperatureToGenericColorName colorTemperature
+        utils_sendEvent name:'colorName', value:colorName, descriptionText:"Color name is ${colorName}", type:type
+    }
+
+    utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ColorTemperatureMireds=${mireds} (${temperature}K), ${colorName}), ColorMode=${colorMode}"
 }
 
 // Implementation for capability.Brightness
@@ -622,32 +655,22 @@ void parse(String description) {
         // Report/Read Attributes Reponse: ColorTemperatureMireds
         case { contains it, [clusterInt:0x0300, commandInt:0x0A, attrInt:0x0007] }:
         case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x0007] }:
-            Integer mireds = Integer.parseInt "${msg.value}", 16
-            Integer colorTemperature = Math.round(1000000 / mireds)
-            String colorName = convertTemperatureToGenericColorName colorTemperature
-            utils_sendEvent name:'colorTemperature', value:colorTemperature, descriptionText:"Color temperature is ${colorTemperature}K", type:'digital'
-            utils_sendEvent name:'colorName', value:colorName, descriptionText:"Color name is ${colorName}", type:'digital'
-            utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ColorTemperatureMireds=${msg.value}(${mireds} mireds, ${colorTemperature}K, ${colorName})"
-            return
         
         // Report/Read Attributes Reponse: ColorMode
         case { contains it, [clusterInt:0x0300, commandInt:0x0A, attrInt:0x0008] }:
         case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x0008] }:
-            String colorMode = msg.value == '02' ? 'CT' : 'RGB'
-            utils_sendEvent name:'colorMode', value:colorMode, descriptionText:"Color mode is ${colorMode}", type:'digital'
-            utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ColorMode=${msg.value}"
+        
+        // Report/Read Attributes Reponse: EnhancedColorMode
+        case { contains it, [clusterInt:0x0300, commandInt:0x0A, attrInt:0x4001] }:
+        case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x4001] }:
+            processMultipleColorTemperatureAttributes msg, type
             return
         
-        // Read Attributes Reponse: ColorTemperaturePhysicalMinMireds
+        // Read Attributes Reponse: ColorTemperaturePhysicalMinMireds, ColorTemperaturePhysicalMaxMireds
         case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x400B] }:
-            state.minMireds = Integer.parseInt "${msg.value}", 16
-            utils_processedZclMessage 'Read Attributes Response', "ColorTemperaturePhysicalMinMireds=${msg.value} (${state.minMireds} mireds, ${Math.round(1000000 / state.maxMireds)}K)"
-            return
-        
-        // Read Attributes Reponse: ColorTemperaturePhysicalMaxMireds
-        case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x400C] }:
-            state.maxMireds = Integer.parseInt "${msg.value}", 16
-            utils_processedZclMessage 'Read Attributes Response', "ColorTemperaturePhysicalMaxMireds=${msg.value} (${state.maxMireds} mireds, ${Math.round(1000000 / state.maxMireds)}K)"
+            state.minMireds = Integer.parseInt msg.value, 16
+            msg.additionalAttrs?.each { if (it.attrId == '400C') state.maxMireds = Integer.parseInt it.value, 16 }
+            utils_processedZclMessage 'Read Attributes Response', "ColorTemperaturePhysicalMinMireds=${msg.value} (${state.minMireds} mireds, ${Math.round(1000000 / state.minMireds)}K), ColorTemperaturePhysicalMaxMireds=${msg.value} (${state.maxMireds} mireds, ${Math.round(1000000 / state.maxMireds)}K)"
             return
         
         // Other events that we expect but are not usefull for capability.ColorTemperature behavior
@@ -663,7 +686,7 @@ void parse(String description) {
         // Report/Read Attributes Reponse: CurrentLevel
         case { contains it, [clusterInt:0x0008, commandInt:0x0A, attrInt:0x0000] }:
         case { contains it, [clusterInt:0x0008, commandInt:0x01, attrInt:0x0000] }:
-            Integer level = msg.value == '00' ? 0 : Math.ceil(Integer.parseInt(msg.value, 16) * 100 / 254)
+            Integer level = msg.value == '00' ? 0 : Math.ceil(Integer.parseInt(msg.value, 16) / 2.54)
             utils_sendEvent name:'level', value:level, descriptionText:"Brightness is ${level}%", type:'digital'
             utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "CurrentLevel=${msg.value} (${level}%)"
             return