+++ ikea-zigbee-drivers/src/capabilities/ColorTemperature.groovy
@@ -70,6 +70,41 @@ void shiftColorTemperature(String direction) {
     String payload = "${utils_payload mode, 2} ${utils_payload stepSize, 4} 0000 ${utils_payload state.minMireds, 4} ${utils_payload state.maxMireds, 4} 00 00"
     utils_sendZigbeeCommands(["he raw 0x${device.deviceNetworkId} 0x01 0x${device.endpointId} 0x0300 {11434C ${payload}}"])
 }
+private void processMultipleColorTemperatureAttributes(Map msg, String type) {
+    Map<Integer, String> attributes = [:]
+    attributes[msg.attrInt] = msg.value
+    msg.additionalAttrs?.each { attributes[Integer.parseInt(it.attrId, 16)] = it.value }
+
+    Integer mireds = -1
+    Integer temperature = -1
+    String colorMode = null
+    String colorName = null
+    attributes.each {
+        switch (it.key) {
+            case 0x0007:
+                mireds = Integer.parseInt it.value, 16
+                temperature = Math.round(1000000 / mireds)
+                break
+
+            case 0x0008:
+            case 0x4001:
+                colorMode = it.value == '02' ? 'CT' : 'RGB'
+                utils_sendEvent name:'colorMode', value:colorMode, descriptionText:"Color mode is ${colorMode}", type:type
+                break
+        }
+    }
+
+    if (temperature >= 0) utils_sendEvent name:'colorTemperature', value:temperature, descriptionText:"Color temperature is ${temperature}K", type:type
+
+    // Update colorName, if the case
+    if ("${colorMode ?: device.currentValue('colorMode', true)}" == 'CT') {
+        Integer colorTemperature = temperature >= 0 ? temperature : device.currentValue('colorTemperature', true)
+        colorName = convertTemperatureToGenericColorName colorTemperature
+        utils_sendEvent name:'colorName', value:colorName, descriptionText:"Color name is ${colorName}", type:type
+    }
+
+    utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ColorTemperatureMireds=${mireds} (${temperature}K), ${colorName}), ColorMode=${colorMode}"
+}
 {{/ @implementation }}
 {{!--------------------------------------------------------------------------}}
 {{# @updated }}
@@ -103,10 +138,8 @@ state.maxMireds = 600  // Will be updated in refresh()
 {{# @refresh }}
 
 // Refresh for capability.ColorTemperature
-cmds += zigbee.readAttribute(0x0300, 0x0008) // ColorMode
-cmds += zigbee.readAttribute(0x0300, 0x0007) // ColorTemperatureMireds
-cmds += zigbee.readAttribute(0x0300, 0x400B) // ColorTemperaturePhysicalMinMireds
-cmds += zigbee.readAttribute(0x0300, 0x400C) // ColorTemperaturePhysicalMaxMireds
+cmds += zigbee.readAttribute(0x0300, [0x0007, 0x0008]) // ColorTemperatureMireds, ColorMode
+cmds += zigbee.readAttribute(0x0300, [0x400B, 0x400C]) // ColorTemperaturePhysicalMinMireds, ColorTemperaturePhysicalMaxMireds
 {{/ @refresh }}
 {{!--------------------------------------------------------------------------}}
 {{# @events }}
@@ -117,32 +150,22 @@ cmds += zigbee.readAttribute(0x0300, 0x400C) // ColorTemperaturePhysicalMaxMired
 // Report/Read Attributes Reponse: ColorTemperatureMireds
 case { contains it, [clusterInt:0x0300, commandInt:0x0A, attrInt:0x0007] }:
 case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x0007] }:
-    Integer mireds = Integer.parseInt "${msg.value}", 16
-    Integer colorTemperature = Math.round(1000000 / mireds)
-    String colorName = convertTemperatureToGenericColorName colorTemperature
-    utils_sendEvent name:'colorTemperature', value:colorTemperature, descriptionText:"Color temperature is ${colorTemperature}K", type:'digital'
-    utils_sendEvent name:'colorName', value:colorName, descriptionText:"Color name is ${colorName}", type:'digital'
-    utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ColorTemperatureMireds=${msg.value}(${mireds} mireds, ${colorTemperature}K, ${colorName})"
-    return
 
 // Report/Read Attributes Reponse: ColorMode
 case { contains it, [clusterInt:0x0300, commandInt:0x0A, attrInt:0x0008] }:
 case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x0008] }:
-    String colorMode = msg.value == '02' ? 'CT' : 'RGB'
-    utils_sendEvent name:'colorMode', value:colorMode, descriptionText:"Color mode is ${colorMode}", type:'digital'
-    utils_processedZclMessage "${msg.commandInt == 0x0A ? 'Report' : 'Read'} Attributes Response", "ColorMode=${msg.value}"
+
+// Report/Read Attributes Reponse: EnhancedColorMode
+case { contains it, [clusterInt:0x0300, commandInt:0x0A, attrInt:0x4001] }:
+case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x4001] }:
+    processMultipleColorTemperatureAttributes msg, type
     return
 
-// Read Attributes Reponse: ColorTemperaturePhysicalMinMireds
+// Read Attributes Reponse: ColorTemperaturePhysicalMinMireds, ColorTemperaturePhysicalMaxMireds
 case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x400B] }:
-    state.minMireds = Integer.parseInt "${msg.value}", 16
-    utils_processedZclMessage 'Read Attributes Response', "ColorTemperaturePhysicalMinMireds=${msg.value} (${state.minMireds} mireds, ${Math.round(1000000 / state.maxMireds)}K)"
-    return
-
-// Read Attributes Reponse: ColorTemperaturePhysicalMaxMireds
-case { contains it, [clusterInt:0x0300, commandInt:0x01, attrInt:0x400C] }:
-    state.maxMireds = Integer.parseInt "${msg.value}", 16
-    utils_processedZclMessage 'Read Attributes Response', "ColorTemperaturePhysicalMaxMireds=${msg.value} (${state.maxMireds} mireds, ${Math.round(1000000 / state.maxMireds)}K)"
+    state.minMireds = Integer.parseInt msg.value, 16
+    msg.additionalAttrs?.each { if (it.attrId == '400C') state.maxMireds = Integer.parseInt it.value, 16 }
+    utils_processedZclMessage 'Read Attributes Response', "ColorTemperaturePhysicalMinMireds=${msg.value} (${state.minMireds} mireds, ${Math.round(1000000 / state.minMireds)}K), ColorTemperaturePhysicalMaxMireds=${msg.value} (${state.maxMireds} mireds, ${Math.round(1000000 / state.maxMireds)}K)"
     return
 
 // Other events that we expect but are not usefull for capability.ColorTemperature behavior